Move conftest to root dir
